criado componente sharebutton
